@@ -7,23 +7,24 @@
 #include <Whip.h>
 
 #include <iostream>
+#include <memory>
 
 int main(int argc, char *argv[]) {
-  Beverage *beverage1 = new Espresso();
-  std::cout << beverage1->getDescription() << " $" << beverage1->cost()
+  Espresso beverage1;
+  std::cout << beverage1.getDescription() << " $" << beverage1.cost()
             << std::endl;
 
-  Beverage *beverage2 = new DarkRoast();
-  beverage2 = new Mocha(beverage2);
-  beverage2 = new Mocha(beverage2);
-  beverage2 = new Whip(beverage2);
+  std::unique_ptr<Beverage> beverage2 = std::make_unique<DarkRoast>();
+  beverage2 = std::make_unique<Mocha>(std::move(beverage2));
+  beverage2 = std::make_unique<Mocha>(std::move(beverage2));
+  beverage2 = std::make_unique<Whip>(std::move(beverage2));
   std::cout << beverage2->getDescription() << " $" << beverage2->cost()
             << std::endl;
 
-  Beverage *beverage3 = new HouseBlend();
-  beverage3 = new Soy(beverage3);
-  beverage3 = new Mocha(beverage3);
-  beverage3 = new Whip(beverage3);
+  std::unique_ptr<Beverage> beverage3 = std::make_unique<HouseBlend>();
+  beverage3 = std::make_unique<Soy>(std::move(beverage3));
+  beverage3 = std::make_unique<Mocha>(std::move(beverage3));
+  beverage3 = std::make_unique<Whip>(std::move(beverage3));
   std::cout << beverage3->getDescription() << " $" << beverage3->cost()
             << std::endl;
   return 0;
