@@ -5,8 +5,8 @@
 
 class Espresso : public Beverage {
 public:
-  Espresso();
-  double cost() override;
+  std::string getDescription() const override;
+  double cost() const override;
 };
 
 #endif
