@@ -5,12 +5,9 @@
 
 class Soy : public CondimentDecorator {
 public:
-  Soy(Beverage *beverage);
-  std::string getDescription() override;
-  double cost() override;
-
-private:
-  Beverage *beverage;
+  Soy(std::unique_ptr<Beverage> beverage);
+  std::string getDescription() const override;
+  double cost() const override;
 };
 
 #endif
