@@ -5,12 +5,9 @@
 
 class Whip : public CondimentDecorator {
 public:
-  Whip(Beverage *beverage);
-  std::string getDescription() override;
-  double cost() override;
-
-private:
-  Beverage *beverage;
+  Whip(std::unique_ptr<Beverage> beverage);
+  std::string getDescription() const override;
+  double cost() const override;
 };
 
 #endif
