@@ -1,9 +1,10 @@
 #include <Milk.h>
 
-Milk::Milk(Beverage *beverage) { this->beverage = beverage; }
+Milk::Milk(std::unique_ptr<Beverage> beverage)
+    : CondimentDecorator(std::move(beverage)) {}
 
-std::string Milk::getDescription() {
+std::string Milk::getDescription() const {
   return beverage->getDescription() + ", Milk";
 }
 
-double Milk::cost() { return beverage->cost() + 0.10; }
+double Milk::cost() const { return beverage->cost() + 0.10; }
