@@ -1,9 +1,10 @@
 #include <Mocha.h>
 
-Mocha::Mocha(Beverage *beverage) { this->beverage = beverage; }
+Mocha::Mocha(std::unique_ptr<Beverage> beverage)
+    : CondimentDecorator(std::move(beverage)) {}
 
-std::string Mocha::getDescription() {
+std::string Mocha::getDescription() const {
   return beverage->getDescription() + ", Mocha";
 }
 
-double Mocha::cost() { return beverage->cost() + 0.20; }
+double Mocha::cost() const { return beverage->cost() + 0.20; }
