@@ -1,9 +1,11 @@
 #include <Soy.h>
 
-Soy::Soy(Beverage *beverage) { this->beverage = beverage; }
+Soy::Soy(std::unique_ptr<Beverage> beverage)
+    : CondimentDecorator(std::move(beverage)) { 
+    }
 
-std::string Soy::getDescription() {
+std::string Soy::getDescription() const {
   return beverage->getDescription() + ", Soy";
 }
 
-double Soy::cost() { return beverage->cost() + 0.15; }
+double Soy::cost() const { return beverage->cost() + 0.15; }
